docs(examples): add camel case plugin example
