+++ examples/basic/package.json
