+++ examples/camel-case/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "example",
+  "scripts": {
+    "dev": "ts-node --transpile-only index.ts"
+  },
+  "dependencies": {
+    "@prisma/client": "^5.0.0"
+  },
+  "devDependencies": {
+    "prisma": "^5.0.0",
+    "prisma-kysely": "^1.7.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.0"
+  }
+}
